pyproject.toml
[tool.poetry]
name = "rag-pipeline"
version = "0.1.0"
description = "RAG Pipeline using LlamaIndex"
authors = ["Michele Makiki <michele.makiki@gmail.com>"]
packages = [
    { include = "rag_pipeline" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
llama-index = "^0.10.0"
llama-index-llms-openai = "^0.1.6"
llama-index-embeddings-openai = "^0.1.6"
llama-index-readers-file = "^0.1.6"
llama-index-vector-stores-chroma = "^0.1.3"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
chromadb = "^0.4.24"
tiktoken = "^0.6.0"
pypdf = "^4.0.2"
docx2txt = "^0.8"
beautifulsoup4 = "^4.12.3"
onnxruntime = "1.16.3"
python-docx = "^1.1.0"
fpdf = "^1.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"